Change the slot of mByEventBut and mByPhasesBut, because setAutoExclusive, let uncheck both. In that case, we are in a uncertain state. Modify mHPDEdit and mFFTLenCombo and MHFactorEdit form to a rounded rectangle. Delete ResultsView::resizeEvent because it is done naturally with ResultsView::paintEvent try to optimize repaint of ResultsView according to mByEventBut and mByPhasesBut fix the showing of the maximum Y on the axis, before the maximum was set by max*10% In GraphViewPhase moving code from updateLayout to PaintEvent

+++ src/mcmc/MHVariable.cpp
@@ -74,7 +74,7 @@ QVector<double> MHVariable::acceptationForChain(const QList<Chain>& chains, int 
     
     for(int i=0; i<chains.size(); ++i)
     {
-        unsigned long long chainSize = chains[i].mNumBurnIter + (chains[i].mBatchIndex * chains[i].mNumBatchIter) + chains[i].mNumRunIter / chains[i].mThinningInterval;
+       int chainSize = chains[i].mNumBurnIter + (chains[i].mBatchIndex * chains[i].mNumBatchIter) + chains[i].mNumRunIter / chains[i].mThinningInterval;
         
         if(i == index)
         {
